+++ testutil/simtesting/utils.go
@@ -36,6 +36,7 @@ func SendMsg(
 
 	txGen := simappparams.MakeTestEncodingConfig().TxConfig
 	tx, err := helpers.GenTx(
+		r,
 		txGen,
 		[]sdk.Msg{msg},
 		fees,